+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Our versioning strategy is as follows:
 - Minor: non-breaking feature additions – no breaking changes (e.g. new features, improvements)
 - Major: new features + breaking changes (e.g. framework upgrades, major architectural changes, major features)
 
-## unreleased
+## 2.1.0
 
 ### 🎉 New Features & Improvements
 
@@ -31,7 +31,6 @@ Our versioning strategy is as follows:
 * `[core] [content]` Fix GraphQL client factory ignoring custom `fetch` and related options ([#418](https://github.com/Sitecore/content-sdk/pull/418))
 * `[nextjs]` AppRouter - NextLink is throwing Locale error in dev mode ([#427](https://github.com/Sitecore/content-sdk/pull/427))
 * `[react]` Form component loses interactivity and state between rerenders ([#447](https://github.com/Sitecore/content-sdk/pull/447))
-* `[nextjs]` BYOC and FEAAS are broken when client component map generation is disabled ([#467](https://github.com/Sitecore/content-sdk/pull/467))
 
 ### ✨ Chores